// app/api/types.gen.ts
export type Config = {
    server: ServerConfig;
    logging: LoggingConfig;
};
// ...
export type LoggingConfig = {
    /**
     * Level
     */
    level: 'debug' | 'info' | 'warn' | 'error';
    /**
     * File
     */
    file: string;
};

/**
 * ServerConfig
 */
export type ServerConfig = {
    /**
     * Host
     */
    host: string;
    /**
     * Port
     */
    port: number;
    /**
     * Use Ssl
     */
    use_ssl: boolean;
};